Update GeoIP.jl
Change name of numericize to octet_to_int, add int_to_octet and add both functions to export.

The octet_to_int function really isn't needed by me, but I regularly need to use the int_to_octet function to translate Int's stored in a database to give people back the octet. It feels like this is the best place if any to have this function available in Julia.